DiscIO: Move some code from DiscExtractor to new file DiscUtils
