+++ Source/Core/DolphinQt/Config/FilesystemWidget.cpp
@@ -18,6 +18,7 @@
 #include <future>
 
 #include "DiscIO/DiscExtractor.h"
+#include "DiscIO/DiscUtils.h"
 #include "DiscIO/Filesystem.h"
 #include "DiscIO/Volume.h"
